feat(slider): add improvements, optimisations and some features
 - simplify file structure of slider
 - minor renaming and refactoring
 - add hover, tab key focus, active states
 - add variants features: sizes, colors
