--- src/components/Slider/PlainSlider/PlainSlider.scss
@@ -1,23 +1,0 @@
-@import "../../../styles/global-css-settings";
-@import "../../../styles/themes";
-
-.monday-slider {
-  &__plain {
-    display: flex;
-    touch-action: none;
-    -webkit-tap-highlight-color: transparent;
-    width: 100%;
-
-    &--small {
-      padding: 0 8px;
-    }
-
-    &--medium {
-      padding: 12px;
-    }
-
-    &--large {
-      padding: 12px;
-    }
-  }
-}
