--- src/components/Slider/PlainSlider/SliderFilledTrack.scss
@@ -1,13 +1,0 @@
-@import "../../../styles/global-css-settings";
-@import "../../../styles/themes";
-
-.monday-slider {
-  &__filled-track {
-    display: block;
-    position: absolute;
-    border-radius: inherit;
-    @include theme-prop(background-color, primary-color);
-    width: 100%;
-    height: inherit;
-  }
-}
