--- src/components/Slider/PlainSlider/SliderRail.scss
@@ -1,32 +1,0 @@
-@import "../../../styles/global-css-settings";
-@import "../../../styles/themes";
-
-.monday-slider {
-  &__rail {
-    box-sizing: content-box;
-    display: inline-block;
-    position: relative;
-    cursor: pointer;
-    touch-action: none;
-    -webkit-tap-highlight-color: transparent;
-    width: 100%;
-
-    &--small {
-      border-radius: 2px;
-      height: 2px;
-      padding: 6px 0;
-    }
-
-    &--medium {
-      border-radius: 2px;
-      height: 4px;
-      padding: 6px 0;
-    }
-
-    &--large {
-      border-radius: 8px;
-      height: 8px;
-      padding: 10px 0;
-    }
-  }
-}
