--- src/components/Slider/PlainSlider/SliderThumb.scss
@@ -1,33 +1,0 @@
-@import "../../../styles/global-css-settings";
-@import "../../../styles/themes";
-
-.monday-slider {
-  &__thumb {
-    position: absolute;
-    border: 3px solid;
-    border-radius: 50%;
-    @include theme-prop(border-color, primary-color);
-    @include theme-prop(background-color, primary-background-color);
-    top: 50%;
-    transform: translate(-50%, -50%);
-
-    &:hover {
-      @include theme-prop(border-color, primary-hover-color);
-    }
-
-    &--small {
-      height: 16px;
-      width: 16px;
-    }
-
-    &--medium {
-      height: 24px;
-      width: 24px;
-    }
-
-    &--large {
-      height: 24px;
-      width: 24px;
-    }
-  }
-}
