+++ src/components/Slider/SliderBase/SliderBase.scss
@@ -1,0 +1,190 @@
+@import "../../../styles/global-css-settings";
+@import "../../../styles/themes";
+
+.monday-slider {
+  &__plain {
+    display: flex;
+    touch-action: none;
+    -webkit-tap-highlight-color: transparent;
+    width: 100%;
+  }
+
+  &__rail {
+    cursor: pointer;
+    display: inline-block;
+    position: relative;
+    touch-action: none;
+    -webkit-tap-highlight-color: transparent;
+    width: 100%;
+  }
+
+  &__track {
+    position: absolute;
+    border-radius: inherit;
+    @include theme-prop(background-color, primary-selected-color);
+    height: inherit;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 100%;
+  }
+
+  &__filled-track {
+    position: absolute;
+    border-radius: inherit;
+    @include theme-prop(background-color, primary-color);
+    height: inherit;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 100%;
+  }
+
+  &__thumb {
+    position: absolute;
+    border: 3px solid;
+    border-radius: 50%;
+    @include theme-prop(border-color, primary-color);
+    @include theme-prop(background-color, primary-background-color);
+    top: 50%;
+    transform: translate(-50%, -50%);
+  }
+
+  // thumb (for NOT-disabled)
+  &:not(&--disabled) {
+    .monday-slider {
+      &__thumb {
+        /* TODO: add fallback for Safari/IE11 */
+        &:focus-visible {
+          @include theme-prop(border-color, primary-hover-color);
+          outline: 3px solid;
+          @include theme-prop(outline-color, primary-selected-color);
+          transform: translate(-50%, -50%) scale(1.33, 1.33);
+          transform-origin: center center;
+          transition: transform 0.2s $expand-animation-timing;
+        }
+
+        &:hover {
+          @include theme-prop(border-color, primary-hover-color);
+          cursor: grab;
+        }
+
+        &:active {
+          cursor: grabbing;
+          transform: translate(-50%, -50%) scale(1.33, 1.33);
+          transform-origin: center center;
+          transition: transform 0.2s $expand-animation-timing;
+        }
+      }
+    }
+  }
+
+  // modifications
+
+  &__plain {
+    &--disabled {
+      * {
+        pointer-events: none;
+      }
+      cursor: not-allowed !important;
+      opacity: 0.4;
+    }
+
+    &--small {
+      padding: 0 8px;
+      .monday-slider {
+        &__rail {
+          border-radius: 2px;
+          height: 2px;
+          padding: 8px 0;
+        }
+        &__thumb {
+          height: 16px;
+          width: 16px;
+        }
+      }
+    }
+
+    &--medium {
+      padding: 12px;
+      .monday-slider {
+        &__rail {
+          border-radius: 2px;
+          height: 4px;
+          padding: 12px 0;
+        }
+        &__thumb {
+          height: 24px;
+          width: 24px;
+        }
+      }
+    }
+
+    &--large {
+      padding: 12px;
+      .monday-slider {
+        &__rail {
+          border-radius: 8px;
+          height: 8px;
+          padding: 10px 0;
+        }
+        &__thumb {
+          height: 24px;
+          width: 24px;
+        }
+      }
+    }
+
+    &--positive {
+      .monday-slider {
+        &__track {
+          @include theme-prop(background-color, positive-color-selected);
+        }
+        &__filled-track {
+          @include theme-prop(background-color, positive-color);
+        }
+        &__thumb {
+          @include theme-prop(border-color, positive-color);
+        }
+      }
+      &:not(&--disabled) {
+        .monday-slider {
+          &__thumb {
+            &:focus-visible {
+              @include theme-prop(border-color, positive-color-hover);
+              @include theme-prop(outline-color, positive-color-selected);
+            }
+            &:hover {
+              @include theme-prop(border-color, positive-color-hover);
+            }
+          }
+        }
+      }
+    }
+
+    &--negative {
+      .monday-slider {
+        &__track {
+          @include theme-prop(background-color, negative-color-selected);
+        }
+        &__filled-track {
+          @include theme-prop(background-color, negative-color);
+        }
+        &__thumb {
+          @include theme-prop(border-color, negative-color);
+        }
+      }
+      &:not(&--disabled) {
+        .monday-slider {
+          &__thumb {
+            &:focus-visible {
+              @include theme-prop(border-color, negative-color-hover);
+              @include theme-prop(outline-color, negative-color-selected);
+            }
+            &:hover {
+              @include theme-prop(border-color, negative-color-hover);
+            }
+          }
+        }
+      }
+    }
+  }
+}
