+++ src/components/Slider/SliderBase/index.js
@@ -1,0 +1,2 @@
+export { default as SliderBase } from "./SliderBase";
+export { useSliderValues } from "./SliderBaseHooks";
