+++ src/components/Slider/SliderBlock.jsx
@@ -1,18 +1,15 @@
 import React, { forwardRef, useRef } from "react";
 import PropTypes from "prop-types";
 import useMergeRefs from "../../hooks/useMergeRefs";
-import { createBemBlockHelper } from "../../helpers/bem-helper";
-import { COMPONENT_ID } from "./SliderCommons";
 import "./SliderBlock.scss";
+import { bem } from "./SliderCommons";
 
-const bem = createBemBlockHelper(COMPONENT_ID);
-
-const SliderBlock = forwardRef(({ children, className, id }, ref) => {
+const SliderBlock = forwardRef(({ children, className, disabled, id }, ref) => {
   const componentRef = useRef(null);
   const mergedRef = useMergeRefs({ refs: [ref, componentRef] });
   console.log("slider: block", { children, className, id });
   return (
-    <div ref={mergedRef} className={bem("", "", className)} id={id}>
+    <div ref={mergedRef} className={bem("", { disabled }, className)} id={id}>
       {children}
     </div>
   );
@@ -24,6 +21,10 @@ SliderBlock.propTypes = {
    */
   className: PropTypes.string,
   /**
+   * If set to true, Component will be disabled
+   */
+  disabled: PropTypes.bool,
+  /**
    * Attribute `id` to be added to the Component's-Root-Node
    */
   id: PropTypes.string
@@ -31,6 +32,7 @@ SliderBlock.propTypes = {
 
 SliderBlock.defaultProps = {
   className: undefined,
+  disabled: false,
   id: undefined
 };
