+++ src/components/Slider/SliderBlock.scss
@@ -3,4 +3,5 @@
 .monday-slider {
   width: 100%;
   min-width: 100px;
+  max-width: 500px;
 }
