+++ src/components/Slider/SliderCommons.js
@@ -1,14 +1,5 @@
+import { createBemBlockHelper } from "../../helpers/bem-helper";
+
 export const COMPONENT_ID = "slider";
 
-export function ensureValueText(valueText, value, formatter) {
-  if (valueText) {
-    return valueText;
-  }
-  if (typeof value === "undefined") {
-    return undefined;
-  }
-  if (typeof formatter !== "function") {
-    return value;
-  }
-  return formatter(value);
-}
+export const bem = createBemBlockHelper(COMPONENT_ID);
